+++ CHANGES.rst
@@ -1,7 +1,7 @@
 0.8.15 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Fix SQLAlchemy V2 support (https://github.com/sqlalchemy-redshift/sqlalchemy-redshift/pull/293)
 
 
 0.8.14 (2023-04-07)