+++ CHANGES.rst
@@ -1,4 +1,4 @@
-0.8.15 (unreleased)
+1.0.0a1 (unreleased)
 -------------------
 
 - Nothing changed yet.